Added struct for UllBin
